pyproject.toml
[project]
name = "ticket-count"
version = "0.1.0"
description = "This script keep track of the number of the tickets worked."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiofiles>=24.1.0",
    "colorama>=0.4.6",
    "pip>=25.0.1",
    "pylint>=3.3.6",
    "python-docx>=1.1.2",
]


[tool.setuptools]
package-dir = { "" = "src" }
packages.find = { where = ["src"] }


